chore updated map view to use useLocation
